Insert wireframe section and link into design doc
I finished the wireframe on figma and put the link into the design doc. Everyone
will need to make a figma account to access the wireframe. I left comments throughout
the wireframe for suggested flow of the feature. These comments are obviously open for
debate and I suspect we will need a few more features not present in the wireframe. This
serves as a baseline to get us started on the project and headed in the right direction.