Added field to model and entity for image
Added a field to model and entity for image and updated the toModel, fromModel, and
update methods on the entity. Also added the condition field to the entity and made
the necessary conditions to the above methods for this change as well.

Changed the update method on Equipment Entity to throw an error if the equipment_id
of the passed in model does not match that of the entity. This should save us from
potential headaches in the future with mismatched updates of entities from models.