Create equipment service and EquipmentNotFound Exception
Implemented the equipment service with basic
methods for checking equipment in and out. The
service will need to be altered in future sprints
as we create tables to store notes about
individual items and as we implement permissions.

The EquipmentNotFound Exception is needed to raise
an error when equipment is checked in/out that
does not exist in the db.