Edit equipment service to combine existing methods into update method
Methods for checking equipment out/in and updating equipment condition
combined into one update method for ease of use. It will be the
responsibility of the API or frontend to edit a model with needed
changes before sending to the service and ultimately updating the db.

methods were also added for inserting equipment into the db and deleting
equipment from the db as deletion will be necessary when equipment is no
longer usable and new equipment will periodically need to be added.