Create equipment service and 'EquipmentNotFoundException' (#17)
* Create equipment service and EquipmentNotFound Exception

Implemented the equipment service with basic
methods for checking equipment in and out. The
service will need to be altered in future sprints
as we create tables to store notes about
individual items and as we implement permissions.

The EquipmentNotFound Exception is needed to raise
an error when equipment is checked in/out that
does not exist in the db.

* Add 'EquipmentService' method for updating condition

* Edit equipment service to combine existing methods into update method

Methods for checking equipment out/in and updating equipment condition
combined into one update method for ease of use. It will be the
responsibility of the API or frontend to edit a model with needed
changes before sending to the service and ultimately updating the db.

methods were also added for inserting equipment into the db and deleting
equipment from the db as deletion will be necessary when equipment is no
longer usable and new equipment will periodically need to be added.