Create waiver comonent, lazy load equipment module
Created the waiver component and can now route to it from the user-equipment component.

Equipment module and subsequent components are now lazy loaded. Getting an error in the console related to this, however, I believe it is because img_url we are recieving from the backend is "placeholder".

Waiver component is still empty so this needs to get done.

Additionally, no longer have to export angular material from equipment module.