34 subtask create and route to waiver component (#45)
* Create waiver comonent, lazy load equipment module

Created the waiver component and can now route to it from the user-equipment component.

Equipment module and subsequent components are now lazy loaded. Getting an error in the console related to this, however, I believe it is because img_url we are recieving from the backend is "placeholder".

Waiver component is still empty so this needs to get done.

Additionally, no longer have to export angular material from equipment module.

* Add signature field to waiver component

Added a signature field that checks if there is input in the field. Cannot submit without input.

Added submit button for the form.

TODO:
- Console is saying form cancelled because it is not connected. need to address this
- Need to route to checkout complete page once this is create.
- need to add actual waiver to the component
- can add checkbox like on wireframe if we want