Add columns to UserEntity and EquipmentEntity
Added column to UserEntity to track if the user has signed the wavier
for checking out equipment. The default value is set to false so that
users who have never checked out equipment are required to sign the
wavier on their first checkout.

Added columns to EquipmentEntity to hold checkout history and condition
notes. the condition notes will be added when a user checks equipment back
in by the ambassador handling the checkin. the checkout history will
automatically be appended to whenever a user completes a checkout

This commit also contains a bug fix within the 'equipment_data' test file
where permissions for updating equipment are added to the ambassador role.