write tests and methods for equipment checkouts
* get_equipment_by_id():
given an equipment id, finds and returns matching equipment item

* create_checkout():
added functionality to update the is_checked_out field of the equipment
item being checked out.

* added new test data for equipment checkouts and permissions

* Wrote tests for new service methods.
note, commented out enforce.assert_called_with() because it was giving
an error because multiple permissions are needed for create_checkout(),
since it calls other methods.
We should either take this out of the tests or make all the permissions
required for these methods the same.