Update equipment permissions actions (#120)
Changed equipment service methods that require permissions to have one
of two actions. If the service method inserts, updates, or deletes an
entity from the database then it gets the 'equipment.crud.checkout'
action otherwise if the service method only retrieves from the database
it gets the 'equipment.view.checkout' action.

Some service methods had to be altered to fit this new permission
scheme. More specifically, the permissions requirements were removed
from the get_equipment_for_request service method. This is because this
method is simply a helper function for other service methods and does
not have a frontend facing API route attached to it.

An __init__.py file was also added to the equipment testing directory so
that it would be recognized in the module tree, and furthermore so that
the imports would function as expected and the pytests could be properly
configured.